# nb-api/api-examples/go-server/go/model_value.go
/*
 * LwM2M NB APIs
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi


import (
	"time"
)



type Value struct {

	Value map[string]interface{} `json:"value,omitempty"`

	ValueType string `json:"valueType,omitempty"`

	DeviceTimestamp time.Time `json:"deviceTimestamp,omitempty"`

	ServerTimestamp time.Time `json:"serverTimestamp,omitempty"`
}

// AssertValueRequired checks if the required fields are not zero-ed
func AssertValueRequired(obj Value) error {
	return nil
}

// AssertValueConstraints checks if the values respects the defined constraints
func AssertValueConstraints(obj Value) error {
	return nil
}