# nb-api/api-examples/mu-server-2/go/api_device.go
/*
 * LwM2M NB APIs
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	//"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// DeviceAPIController binds http requests to an api service and writes the service results to the http response
type DeviceAPIController struct {
	service      DeviceAPIServicer
	errorHandler ErrorHandler
}

// DeviceAPIOption for how the controller is set up.
type DeviceAPIOption func(*DeviceAPIController)

// WithDeviceAPIErrorHandler inject ErrorHandler into controller
func WithDeviceAPIErrorHandler(h ErrorHandler) DeviceAPIOption {
	return func(c *DeviceAPIController) {
		c.errorHandler = h
	}
}

// NewDeviceAPIController creates a default api controller
func NewDeviceAPIController(s DeviceAPIServicer, opts ...DeviceAPIOption) Router {
	controller := &DeviceAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the DeviceAPIController
func (c *DeviceAPIController) Routes() Routes {
	return Routes{
		"GetDevice": Route{
			strings.ToUpper("Get"),
			"/lwm2m/v1/devices/{deviceId}",
			c.GetDevice,
		},
	}
}

// GetDevice - Get device information
func (c *DeviceAPIController) GetDevice(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	deviceID := params["deviceId"]

	// Retrieve the device information from the mockDevices map
	device, ok := mockDevices[deviceID]
	if !ok {
		// If the device is not found, return a 404 Not Found error
		c.errorHandler(w, r, NotFoundError{"Device not found"}, nil)
		return
	}

	// Create the response payload
	response := Device{
		EndpointName: device.EndpointName,
		Binding:      device.Binding,
		Registration: device.Registration,
		LwM2MVersion: device.LwM2MVersion,
		Lifetime:     device.Lifetime,
		Operations:   device.Operations,
		Observations: device.Observations,
		Objects:      device.Objects,
	}

	// Encode the response as JSON and write it to the response writer
	EncodeJSONResponse(response, http.StatusOK, w)
}