# nb-api/api-examples/mu-server-2/go/api_device_observation.go
/*
 * LwM2M NB APIs
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// DeviceObservationAPIController binds http requests to an api service and writes the service results to the http response
type DeviceObservationAPIController struct {
	service DeviceObservationAPIServicer
	errorHandler ErrorHandler
}

// DeviceObservationAPIOption for how the controller is set up.
type DeviceObservationAPIOption func(*DeviceObservationAPIController)

// WithDeviceObservationAPIErrorHandler inject ErrorHandler into controller
func WithDeviceObservationAPIErrorHandler(h ErrorHandler) DeviceObservationAPIOption {
	return func(c *DeviceObservationAPIController) {
		c.errorHandler = h
	}
}

// NewDeviceObservationAPIController creates a default api controller
func NewDeviceObservationAPIController(s DeviceObservationAPIServicer, opts ...DeviceObservationAPIOption) Router {
	controller := &DeviceObservationAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the DeviceObservationAPIController
func (c *DeviceObservationAPIController) Routes() Routes {
	return Routes{
		"CreateDeviceObservation": Route{
			strings.ToUpper("Post"),
			"/lwm2m/v1/devices/{deviceId}/observations",
			c.CreateDeviceObservation,
		},
		"DeleteDeviceObservation": Route{
			strings.ToUpper("Delete"),
			"/lwm2m/v1/devices/{deviceId}/observations/{observationId}",
			c.DeleteDeviceObservation,
		},
		"GetDeviceObservation": Route{
			strings.ToUpper("Get"),
			"/lwm2m/v1/devices/{deviceId}/observations/{observationId}",
			c.GetDeviceObservation,
		},
		"GetDeviceObservationNotification": Route{
			strings.ToUpper("Get"),
			"/lwm2m/v1/devices/{deviceId}/observations/{observationId}/notifications/{notificationId}",
			c.GetDeviceObservationNotification,
		},
		"GetDeviceObservationNotifications": Route{
			strings.ToUpper("Get"),
			"/lwm2m/v1/devices/{deviceId}/observations/{observationId}/notifications",
			c.GetDeviceObservationNotifications,
		},
		"GetDeviceObservations": Route{
			strings.ToUpper("Get"),
			"/lwm2m/v1/devices/{deviceId}/observations",
			c.GetDeviceObservations,
		},
	}
}

// CreateDeviceObservation - Create a new device observation
func (c *DeviceObservationAPIController) CreateDeviceObservation(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	deviceIdParam := params["deviceId"]
	if deviceIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"deviceId"}, nil)
		return
	}
	createObservationParam := CreateObservation{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&createObservationParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertCreateObservationRequired(createObservationParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertCreateObservationConstraints(createObservationParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.CreateDeviceObservation(r.Context(), deviceIdParam, createObservationParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeleteDeviceObservation - Delete a device observation
func (c *DeviceObservationAPIController) DeleteDeviceObservation(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	deviceIdParam := params["deviceId"]
	if deviceIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"deviceId"}, nil)
		return
	}
	observationIdParam := params["observationId"]
	if observationIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"observationId"}, nil)
		return
	}
	result, err := c.service.DeleteDeviceObservation(r.Context(), deviceIdParam, observationIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetDeviceObservation - Get the state of a device observation
func (c *DeviceObservationAPIController) GetDeviceObservation(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	deviceIdParam := params["deviceId"]
	if deviceIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"deviceId"}, nil)
		return
	}
	observationIdParam := params["observationId"]
	if observationIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"observationId"}, nil)
		return
	}
	result, err := c.service.GetDeviceObservation(r.Context(), deviceIdParam, observationIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetDeviceObservationNotification - Get the notification of a device observation
func (c *DeviceObservationAPIController) GetDeviceObservationNotification(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	deviceIdParam := params["deviceId"]
	if deviceIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"deviceId"}, nil)
		return
	}
	observationIdParam := params["observationId"]
	if observationIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"observationId"}, nil)
		return
	}
	notificationIdParam := params["notificationId"]
	if notificationIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"notificationId"}, nil)
		return
	}
	result, err := c.service.GetDeviceObservationNotification(r.Context(), deviceIdParam, observationIdParam, notificationIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetDeviceObservationNotifications - Get the device observation's notifications
func (c *DeviceObservationAPIController) GetDeviceObservationNotifications(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	query := r.URL.Query()
	deviceIdParam := params["deviceId"]
	if deviceIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"deviceId"}, nil)
		return
	}
	observationIdParam := params["observationId"]
	if observationIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"observationId"}, nil)
		return
	}
	pageParam, err := parseNumericParameter[int32](
		query.Get("page"),
		WithParse[int32](parseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	sizeParam, err := parseNumericParameter[int32](
		query.Get("size"),
		WithParse[int32](parseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	result, err := c.service.GetDeviceObservationNotifications(r.Context(), deviceIdParam, observationIdParam, pageParam, sizeParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetDeviceObservations - Get all the device observations
func (c *DeviceObservationAPIController) GetDeviceObservations(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	query := r.URL.Query()
	deviceIdParam := params["deviceId"]
	if deviceIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"deviceId"}, nil)
		return
	}
	pageParam, err := parseNumericParameter[int32](
		query.Get("page"),
		WithParse[int32](parseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	sizeParam, err := parseNumericParameter[int32](
		query.Get("size"),
		WithParse[int32](parseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	result, err := c.service.GetDeviceObservations(r.Context(), deviceIdParam, pageParam, sizeParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)
}