# nb-api/api-examples/mu-server-2/go/model_common_notification.go
/*
 * LwM2M NB APIs
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi




type CommonNotification struct {

	DeviceId string `json:"deviceId,omitempty"`
}

// AssertCommonNotificationRequired checks if the required fields are not zero-ed
func AssertCommonNotificationRequired(obj CommonNotification) error {
	return nil
}

// AssertCommonNotificationConstraints checks if the values respects the defined constraints
func AssertCommonNotificationConstraints(obj CommonNotification) error {
	return nil
}