# nb-api/api-examples/mu-server-2/go/model_paginated_observation_notification.go
/*
 * LwM2M NB APIs
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import "time"

type PaginatedObservationNotification struct {
	Type            string    `json:"type"`
	NotificationId  string    `json:"notificationId"`
	DeviceTimestamp time.Time `json:"deviceTimestamp,omitempty"`
	ServerTimestamp time.Time `json:"serverTimestamp"`
}

// AssertPaginatedObservationNotificationRequired checks if the required fields are not zeroed
func AssertPaginatedObservationNotificationRequired(obj PaginatedObservationNotification) error {
	elements := map[string]interface{}{
		"type":            obj.Type,
		"notificationId":  obj.NotificationId,
		"serverTimestamp": obj.ServerTimestamp,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertPaginatedObservationNotificationConstraints checks if the values respect the defined constraints
func AssertPaginatedObservationNotificationConstraints(obj PaginatedObservationNotification) error {
	return nil
}