# nb-api/api-examples/mu-server-2/go/model_pagination.go
/*
 * LwM2M NB APIs
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi




type Pagination struct {

	TotalElements int32 `json:"totalElements,omitempty"`

	TotalPages int32 `json:"totalPages,omitempty"`

	PageSize int32 `json:"pageSize,omitempty"`

	Count int32 `json:"count,omitempty"`

	Page int32 `json:"page,omitempty"`

	NextPage string `json:"nextPage,omitempty"`
}

// AssertPaginationRequired checks if the required fields are not zero-ed
func AssertPaginationRequired(obj Pagination) error {
	return nil
}

// AssertPaginationConstraints checks if the values respects the defined constraints
func AssertPaginationConstraints(obj Pagination) error {
	return nil
}